adds useStyle hook from MUI to useStyles.js file in hooks dir
