+++ src/components/Settings.jsx
@@ -2,6 +2,10 @@ import { useState,useEffect } from 'react'
 
 import Button from '@mui/material/Button';
 import Switch from '@mui/material/Switch';
+import Card from '@mui/material/Card';
+import CardHeader from '@mui/material/CardHeader';
+import CardContent from '@mui/material/CardContent';
+import Container from '@mui/material/Container';
 
 const Settings = () => {
 
@@ -19,23 +23,34 @@ const Settings = () => {
   },[defaults])
 
   return(
-    <div>
-      <h1>Update Your Settings:</h1>
-      <div>
-      <h2>Show Completed Todos</h2>
-      dont show<Switch onChange={()=> setDefaults({...defaults,completedTodos: !defaults.completedTodos}) } value={true} variant="contained"/>show
-      </div> 
-      <div>
-        number of todos per page
-        <Button
-          onClick={() => setDefaults({...defaults,totalTodos: defaults.totalTodos - 1})}
-        > - </Button>
-        {defaults.totalTodos}
-        <Button
-          onClick={() => setDefaults({...defaults,totalTodos: defaults.totalTodos + 1})}
-        > + </Button>
-      </div>
-    </div>
+      <Container 
+          maxWidth="md"
+          align="center"
+
+        >
+          <h1>Adjust Settings</h1>
+          <Card boolean="true">
+            <CardHeader title='Show Completed'></CardHeader>
+            <CardContent>
+              dont show
+              <Switch 
+                onChange={()=> setDefaults({...defaults,completedTodos: !defaults.completedTodos})} 
+                value={true} variant="contained"
+                />show
+            </CardContent>
+          </Card>
+          <Card boolean="true">
+            <CardHeader title='Items per page'></CardHeader>
+              <Button
+                // className={classes.button}
+                onClick={() => setDefaults({...defaults,totalTodos: defaults.totalTodos - 1})}
+              > - </Button>
+              {defaults.totalTodos}
+              <Button
+                onClick={() => setDefaults({...defaults,totalTodos: defaults.totalTodos + 1})}
+              > + </Button>
+          </Card>
+      </Container>
   )
 }
 export default Settings;
